+++ src/main/java/com/monthlyib/server/openapi/user/dto/PasswordResetRequestDto.java
@@ -1,0 +1,21 @@
+package com.monthlyib.server.openapi.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PasswordResetRequestDto {
+
+    private String username;
+
+    private String email;
+
+    private String verifyNum;
+}